requirements.txt
Flask==2.3.5
Flask-Mail==1.0.1
python-dotenv==1.0.1
pandas==2.1.1
scikit-learn==1.3.2
numpy==1.27.5
geopy==2.6.0
gunicorn==21.2.0